backend/user/requirements.txt
djangorestframework~=3.11.1
django-filter~=2.4.0
drf-yasg
django-cors-headers
pymongo==3.11.0
djongo==1.3.3
sqlparse
djoser
common-structure-microservices
Django~=3.0.5
dnspython
exrex~=0.10.5
munch~=2.5.0
rstr
py_eureka_client==0.7.4